TVGUI-17 Handle types through the types sub-directory
