+++ src/Pages/Home.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 
 import TVGuide from "../components/TVGuide";
 import { getGuide } from "../requests";
-import { Guide } from "../utils";
+import { Guide } from "../utils/types";
 
 const Home = () => {
     const [guide, setGuide] = useState<Guide | null>(null);
