+++ src/Pages/RemindersPage.tsx
@@ -5,7 +5,7 @@ import { EditOutlined, DeleteOutlined } from "@ant-design/icons";
 import { AddReminder, EditReminder } from '../components/Reminders';
 import { RemindersContext, UserContext } from '../contexts';
 import { deleteReminder } from '../requests';
-import { Reminder } from '../utils';
+import { Reminder } from '../utils/types';
 import './styles/RemindersPage.css';
