+++ src/components/AddShow/index.tsx
@@ -8,7 +8,7 @@ import { PrevArrow, NextArrow } from "./ArrowComponents";
 import { UserContext } from "../../contexts/UserContext";
 import { addShowToList } from "../../requests";
 import { getShowSeasons, searchNewShow } from "../../requests";
-import { SeasonSearch, ShowSearchResult } from "../../utils/types/index";
+import { SeasonSearch, ShowSearchResult } from "../../utils/types";
 import './AddShow.scss';
 
 interface AddShowProps {
