+++ src/contexts/ErrorsContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
 
-import { ErrorsContextModel } from "../utils";
+import { ErrorsContextModel } from "../utils/types";
 
 export const ErrorsContext = createContext<ErrorsContextModel>(null);
