+++ src/contexts/RecordedShowsContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
 
-import { RecordedShowsContextModel } from "../utils";
+import { RecordedShowsContextModel } from "../utils/types";
 
 export const RecordedShowsContext = createContext<RecordedShowsContextModel>(null);
