+++ src/contexts/SearchListContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
 
-import { SearchListContextModel } from "../utils";
+import { SearchListContextModel } from "../utils/types";
 
 export const SearchListContext = createContext<SearchListContextModel>(null);
