+++ src/contexts/UserContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
 
-import { UserContextModel } from '../utils';
+import { UserContextModel } from '../utils/types';
 
 export const UserContext = createContext<UserContextModel>(null);
