+++ src/requests/index.ts
@@ -15,8 +15,8 @@ import {
     UserResponses,
     SearchItemResponses,
     NewUserDetails,
-} from "../utils";
-import { ShowSearchResult } from "../utils/types/index";
+    ShowSearchResult
+} from "../utils/types";
 
 const baseURL = 'http://127.0.0.1:5000/api';
