+++ src/tests/test_data/index.ts
@@ -9,7 +9,7 @@ import {
     SearchItemResponses,
     User,
     UserResponses
-} from "../../utils";
+} from "../../utils/types";
 
 
 export const guide: Guide = {
